+++ p2panda-auth/src/group/action.rs
@@ -1,10 +1,14 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+#[cfg(any(test, feature = "serde"))]
+use serde::{Deserialize, Serialize};
+
 use crate::Access;
 use crate::group::GroupMember;
 
 /// Actions for creating groups and modifying group membership.
 #[derive(Clone, Debug, PartialEq)]
+#[cfg_attr(any(test, feature = "serde"), derive(Deserialize, Serialize))]
 pub enum GroupAction<ID, C> {
     Create {
         initial_members: Vec<(GroupMember<ID>, Access<C>)>,