+++ p2panda-auth/src/group/member.rs
@@ -1,5 +1,8 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+#[cfg(any(test, feature = "serde"))]
+use serde::{Deserialize, Serialize};
+
 use crate::traits::IdentityHandle;
 
 /// A group member which can be a single individual or another group.
@@ -7,6 +10,7 @@ use crate::traits::IdentityHandle;
 /// The `Group` variant can be used to express nested group relations. In both cases, the member
 /// identifier is the same generic ID.
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, PartialOrd, Ord)]
+#[cfg_attr(any(test, feature = "serde"), derive(Deserialize, Serialize))]
 pub enum GroupMember<ID> {
     Individual(ID),
     Group(ID),